fine

+++ lib/settings/pages/devices.dart
@@ -2,9 +2,14 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_application_1/services/database_service.dart';
 import 'package:flutter_application_1/services/device_model.dart';
+import 'package:flutter_application_1/settings/sharedPreferencesHelper.dart';
 import 'package:flutter_application_1/utils/colors.dart';
 import 'package:flutter_application_1/utils/dimensions.dart';
+import 'package:flutter_application_1/utils/global_variables.dart';
+import 'package:flutter_application_1/widgets/dialog.dart';
+import 'package:flutter_application_1/widgets/dialog_content.dart';
 import 'package:flutter_application_1/widgets/edited_text.dart';
+import 'package:flutter_application_1/widgets/ip_dialog_content.dart';
 import 'package:mask_text_input_formatter/mask_text_input_formatter.dart';
 
 class Devices extends StatefulWidget {
@@ -20,12 +25,25 @@ class _DevicesState extends State<Devices> {
   final DatabaseService _databaseService = DatabaseService.instance;
   int selectedDevice = -1;
   int selectedIndex = 0;
+  int finalDevicesCount = 5;
+  int currentDevicesCount = 0;
+  bool isLoaded = false;
 
   @override
   void initState() {
     super.initState();
     nameController = TextEditingController();
     numController = TextEditingController();
+    WidgetsBinding.instance.addPostFrameCallback((_){
+      _getDb();
+      setState(() {});
+    });
+      
+  }
+
+  _getDb() async {
+    currentDevicesCount = await _databaseService.getCount();
+    isLoaded = true;
   }
 
   @override
@@ -35,17 +53,14 @@ class _DevicesState extends State<Devices> {
     super.dispose();
   }
 
-  int finalDevicesCount = 5;
-  int currentDevicesCount = 0;
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
         body: Padding(
       padding: EdgeInsets.only(left: Dimensions.margin10Width * 2.7),
       child: Column(
-        children: [
-          SizedBox(
+        children: [ isLoaded
+        ? SizedBox(
             height: Dimensions.margin10Height * 10,
             child: Row(
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
@@ -76,15 +91,19 @@ class _DevicesState extends State<Devices> {
                         nameController.text = "";
                         numController.text = "";
                       });
-                      (currentDevicesCount == finalDevicesCount)
-                          ? null
-                          : openDialog();
+                      if (currentDevicesCount == finalDevicesCount){
+                        null;
+                      } else {
+                        openDialog();
+                        Feedback.forTap(context);
+                      }
                     },
                   ),
                 ))
               ],
             ),
-          ),
+          )
+        : SizedBox(height: Dimensions.margin10Height * 10),
           Expanded(
             child: Container(
               padding: EdgeInsets.only(
@@ -101,7 +120,8 @@ class _DevicesState extends State<Devices> {
           )
         ],
       ),
-    ));
+    )
+        );
   }
 
   Widget _devicesList() {
@@ -240,8 +260,8 @@ class _DevicesState extends State<Devices> {
                       onTap: () {
                         setState(() {
                           nameController.text = deviceInfo.name;
-                          numController.text =
-                              ipFormatter.maskText(deviceInfo.serialNum);
+                          numController.text =ipFormatter.maskText(deviceInfo.serialNum);
+                          Feedback.forTap(context);
                         });
                         updateDialog();
                       },
@@ -272,7 +292,7 @@ class _DevicesState extends State<Devices> {
                               width: Dimensions.border1)),
                       child: EditedText(
                           color: AppColors.whiteText,
-                          text: "${deviceInfo.id}",
+                          text: "${deviceInfo.ipNum}",
                           size: Dimensions.font10 * 3.5,
                           fontWeight: FontWeight.w500),
                     ),
@@ -292,7 +312,12 @@ class _DevicesState extends State<Devices> {
                             size: Dimensions.font10 * 3,
                             fontWeight: FontWeight.w500),
                       ),
-                      onTap: () {},
+                      onTap: () {
+                        Feedback.forTap(context);
+                        IPDialog();
+                        setState(() {
+                        });
+                      },
                     )
                   ],
                 ),
@@ -309,16 +334,23 @@ class _DevicesState extends State<Devices> {
                     ),
                     child: EditedText(
                         color: AppColors.blackText,
-                        text: 'Удалить устройство',
+                        text: GlobalVariables.canDelete == true
+                        ? 'Удалить устройство'
+                        : 'Хрен',
                         size: Dimensions.font10 * 3,
                         fontWeight: FontWeight.w500),
                   ),
                   onTap: () {
-                    _databaseService.deleteDevice(deviceInfo.id);
-                    setState(() {
-                      currentDevicesCount--;
-                      selectedIndex = 0;
-                    });
+                      Feedback.forTap(context);
+                      if (GlobalVariables.canDelete == true){
+                      _databaseService.deleteDevice(deviceInfo.id);
+                      setState(() {
+                        currentDevicesCount--;
+                        selectedIndex = 0;
+                      });
+                      } else {
+                        errorDelete();
+                      }
                   },
                 )
               ],
@@ -339,7 +371,7 @@ class _DevicesState extends State<Devices> {
       context: context,
       builder: (context) => AlertDialog(
             backgroundColor: AppColors.darkMainColor,
-            content: dialogContent(),
+            content: DialogContent(nameController: nameController, numController: numController,),
             actions: [
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
@@ -363,6 +395,7 @@ class _DevicesState extends State<Devices> {
                           nameController.text.isNotEmpty && numController.text.length == 15) {
                         _databaseService.addDevice(
                             nameController.text, numController.text);
+                            Feedback.forTap(context);
                         setState(() {
                           nameController.text = "";
                           numController.text = "";
@@ -377,12 +410,80 @@ class _DevicesState extends State<Devices> {
             ],
           ));
 
+  Future<void> IPDialog() => showDialog(
+    barrierDismissible: false,
+    context: context,
+    builder: (context) {
+      return AlertDialog(
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(Dimensions.cornerRadius25)),
+      backgroundColor: AppColors.darkMainColor,
+      contentPadding: EdgeInsets.zero,
+      content: IpDialogContent(),
+      actions: [
+        GestureDetector(
+                    child: Container(
+                        alignment: Alignment.center,
+                        width: Dimensions.margin10Width * 38,
+                        height: Dimensions.margin10Height * 7,
+                        decoration: BoxDecoration(
+                            borderRadius: BorderRadius.circular(
+                                Dimensions.cornerRadius20),
+                            color: AppColors.yellowButtonColor),
+                        child: EditedText(
+                            color: AppColors.blackText,
+                            text: 'Сохранить',
+                            size: Dimensions.font10 * 3.3,
+                            fontWeight: FontWeight.w500)),
+                    onTap: () {
+                      if (GlobalVariables.selectedIP != -1) {
+                        setState(() {
+                          _databaseService.addIPDevice(selectedDevice, GlobalVariables.selectedIP_text);
+                          GlobalVariables.changeIP(-1);
+                      });
+                        Navigator.of(context).pop();
+                      }
+                    },
+                  ),
+      ],
+    );
+        }  
+    );
+        
+    Future<void> errorDelete() => showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+            backgroundColor: AppColors.darkMainColor,
+            content: Container(
+              alignment: Alignment.center,
+              width: Dimensions.margin10Width*100,
+              height: Dimensions.margin10Height*35,
+              child: EditedText(color: AppColors.whiteText, text: 'Сначала отключите все таймеры', size: Dimensions.font10*5, fontWeight: FontWeight.w700),
+              ),
+            actions: [
+              GestureDetector(
+                    child: Container(
+                        alignment: Alignment.center,
+                        width: Dimensions.margin10Width * 25,
+                        height: Dimensions.margin10Height * 9,
+                        child: EditedText(
+                            color: AppColors.whiteText,
+                            text: 'ОК',
+                            size: Dimensions.font10 * 4,
+                            fontWeight: FontWeight.w600)),
+                    onTap: () {
+                        Feedback.forTap(context);
+                        Navigator.of(context).pop();
+                    },
+                  ),
+            ],
+          ));
+  
   Future<void> updateDialog() => showDialog(
       barrierDismissible: false,
       context: context,
       builder: (context) => AlertDialog(
             backgroundColor: AppColors.darkMainColor,
-            content: dialogContent(),
+            content: DialogContent(nameController: nameController, numController: numController,),
             actions: [
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
@@ -406,7 +507,6 @@ class _DevicesState extends State<Devices> {
                           nameController.text.isNotEmpty && numController.text.length == 15) {
                         _databaseService.updateDevice(selectedDevice,
                             nameController.text, numController.text);
-                        print(selectedDevice);
                         setState(() {
                           nameController.text = "";
                           numController.text = "";
@@ -420,125 +520,6 @@ class _DevicesState extends State<Devices> {
             ],
           ));
 
-  Widget dialogContent() {
-    return SizedBox(
-      width: Dimensions.margin10Width * 125,
-      height: Dimensions.margin10Height * 90,
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.spaceAround,
-        children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.end,
-            children: [
-              GestureDetector(
-                child: Container(
-                  width: Dimensions.margin10Width * 9.2,
-                  height: Dimensions.margin10Height * 6.5,
-                  decoration: BoxDecoration(
-                      borderRadius:
-                          BorderRadius.circular(Dimensions.cornerRadius15),
-                      color: AppColors.blueButtonColor),
-                  child: Icon(Icons.close,
-                      color: AppColors.whiteText, size: Dimensions.font10 * 4),
-                ),
-                onTap: () => Navigator.of(context).pop(),
-              ),
-            ],
-          ),
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              EditedText(
-                  color: Theme.of(context).colorScheme.tertiary,
-                  text: "Название устройства",
-                  size: Dimensions.font10 * 3.5,
-                  fontWeight: FontWeight.w700),
-              Container(
-                width: Dimensions.margin10Width * 131.6,
-                height: Dimensions.margin10Height * 10.8,
-                child: TextFormField(
-                  controller: nameController,
-                  autofocus: true,
-                  style: TextStyle(
-                      color: Theme.of(context).colorScheme.tertiaryContainer,
-                      fontSize: Dimensions.font10 * 3.5,
-                      fontFamily: 'Roboto',
-                      fontWeight: FontWeight.w500),
-                  decoration: InputDecoration(
-                    filled: true,
-                    fillColor: Theme.of(context).colorScheme.primaryContainer,
-                    border: OutlineInputBorder(
-                        borderRadius:
-                            BorderRadius.circular(Dimensions.cornerRadius15),
-                        borderSide: BorderSide(
-                            color: AppColors.darkerGreyText,
-                            width: Dimensions.border1)),
-                    enabledBorder: OutlineInputBorder(
-                        borderRadius:
-                            BorderRadius.circular(Dimensions.cornerRadius15),
-                        borderSide: BorderSide(
-                            color: AppColors.darkerGreyText,
-                            width: Dimensions.border1)),
-                    focusedBorder: OutlineInputBorder(
-                        borderRadius:
-                            BorderRadius.circular(Dimensions.cornerRadius15),
-                        borderSide: BorderSide(
-                            color: AppColors.darkerGreyText,
-                            width: Dimensions.border1)),
-                  ),
-                ),
-              ),
-            ],
-          ),
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              EditedText(
-                  color: Theme.of(context).colorScheme.tertiary,
-                  text: "Ip устройства",
-                  size: Dimensions.font10 * 3.5,
-                  fontWeight: FontWeight.w700),
-              Container(
-                width: Dimensions.margin10Width * 131.6,
-                height: Dimensions.margin10Height * 10.8,
-                child: TextFormField(
-                  controller: numController,
-                  inputFormatters: [ipFormatter],
-                  style: TextStyle(
-                      color: Theme.of(context).colorScheme.tertiaryContainer,
-                      fontSize: Dimensions.font10 * 3.5,
-                      fontFamily: 'Roboto',
-                      fontWeight: FontWeight.w500),
-                  decoration: InputDecoration(
-                    filled: true,
-                    fillColor: Theme.of(context).colorScheme.primaryContainer,
-                    border: OutlineInputBorder(
-                        borderRadius:
-                            BorderRadius.circular(Dimensions.cornerRadius15),
-                        borderSide: BorderSide(
-                            color: AppColors.darkerGreyText,
-                            width: Dimensions.border1)),
-                    enabledBorder: OutlineInputBorder(
-                        borderRadius:
-                            BorderRadius.circular(Dimensions.cornerRadius15),
-                        borderSide: BorderSide(
-                            color: AppColors.darkerGreyText,
-                            width: Dimensions.border1)),
-                    focusedBorder: OutlineInputBorder(
-                        borderRadius:
-                            BorderRadius.circular(Dimensions.cornerRadius15),
-                        borderSide: BorderSide(
-                            color: AppColors.darkerGreyText,
-                            width: Dimensions.border1)),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
 
   var ipFormatter = new MaskTextInputFormatter(
       mask: '###.###.###.###',
